main/xiaozhi-server/requirements.txt
# Python dependencies - Latest compatible versions (2025)
# Core ML/AI frameworks
torch==2.4.1
torchaudio==2.4.1
numpy>=2.0.2  # Updated for kokoro-tts compatibility (PyTorch 2.4.1 supports NumPy 2.x)
silero_vad==6.0.0

# Web frameworks and networking
websockets==15.0.1
aiohttp==3.12.15
aiohttp_cors==0.7.0
httpx==0.28.1
requests==2.32.5

# Audio processing
pydub==0.25.1
opuslib_next==1.1.2

# AI/LLM APIs
openai==1.109.0  # Constrained by mem0ai requirement: <1.110.0
google-generativeai==0.8.5
elevenlabs==2.16.0
kokoro-tts==2.3.0  # DISABLED: Requires Python 3.9-3.12 (not compatible with Python 3.13+)

# TTS services
edge_tts==7.0.0

# Data processing and serialization
ormsgpack==1.7.0
ruamel.yaml==0.18.15
pyyml==0.0.2

# Logging and utilities
loguru==0.7.3
psutil==7.0.0
portalocker==3.2.0
Jinja2==3.1.6

# Security and authentication
PyJWT==2.8.0
cryptography==46.0.2
PySocks==1.7.1

# Text processing and parsing
chardet==5.2.0
bs4==0.0.2
markitdown==0.1.3

# Console and development tools
aioconsole==0.8.1

# Speech recognition and processing
vosk==0.3.45
openai-whisper==20240930

# Memory and MCP
mem0ai==0.1.118
mcp==1.16.0
mcp-proxy==0.8.2
